Flip around Github::WebhooksController naming
