--- app/controllers/github/webhooks_controller.rb
@@ -1,23 +1,0 @@
-module Github
-  class WebhooksController < ApplicationController
-    skip_before_action :verify_authenticity_token
-    before_action :verify_signature
-
-    respond_to :json
-
-    def push_event
-      render_status :ok
-    end
-
-    private
-
-    def render_status(status)
-      render nothing: true, status: status
-    end
-
-    def verify_signature
-      signature = 'sha1=' + OpenSSL::HMAC.hexdigest(OpenSSL::Digest.new('sha1'), ENV['SECRET_TOKEN'], request.body.read)
-      render_status(:error) unless Rack::Utils.secure_compare(signature, request.headers['HTTP_X_HUB_SIGNATURE'])
-    end
-  end
-end
