+++ app/controllers/webhooks/github_controller.rb
@@ -1,0 +1,9 @@
+module Webhooks
+  class GithubController < WebhookController
+    respond_to :json
+
+    def push_event
+      render_status :ok
+    end
+  end
+end
