+++ spec/routing/webhooks/github_routing_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+describe Webhooks::GithubController do
+  it { expect(post: 'webhooks/github/push_event').to route_to('webhooks/github#push_event') }
+end
